refactor: create LocationUtil for location parsing
